+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - Documentation improvements
+- **Added:** `CAlloc::allocation_usable_size`
 
 ## [0.1.2] - 2021-05-30